mock-sre-lab/app/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
httpx>=0.27,<1
prometheus-client>=0.21,<1
opentelemetry-api>=1.28,<2
opentelemetry-sdk>=1.28,<2
opentelemetry-exporter-otlp-proto-grpc>=1.28,<2
sentry-sdk>=2.14,<3
psutil>=6,<7
pytest>=8,<9

